Limit max pagination depth (#3353)
* Limit max pagination depth

Requesting a resource with a high index number in your paginated
list endpoint results in sql queries having a large offset value.
Such queries get gradually slower the higher the offset is in the
sql query. As an example we saw queries finish in 7ms with an offset of
0. When the offset was 10 million the same query took nearly a minute to
be answered from the database. A user can thus construct quite costly
queries on resouces that have large tables.

This change introduces a limit on how far you can go back in your
requests pagination. The limit is quite high and disabled by default.
It can be enabled via cloud_controller configuration and values below 1
million seem fine. That would mean no page larger than 2000 can be
requested in case of default page size. Ultimately the user is asked to narrow its list
request (by created_at time for example) to reach elements in later pages.

If the limit is disabled it will still produce a warning message in the
cloud controller logs to make operators aware if consumers make queries
that produce large offset values in sql queries. This then just gets a real
issue when then table really has that many elements.


Co-authored-by: Johannes Haass <johannes.haass@sap.com>