Eager load domains for routes associated with a process
The 'routes' association in the 'process' model is used to determine the
'uris' belonging to a process. This also requires the 'name' from the
associated 'domain' which leads to 1 + n SELECT queries (1 * SELECT FROM
routes + n * SELECT FROM domains), where n is the number of associated
routes. When eager loading domains, there are only 2 SELECT queries no
matter how many routes are associated with a process.