Prepare Annotation Key Column Rename
This change introduces views on all annotation tables where the column
key has the alias key_name. The coding then uses these Views and all
occurences of 'key' are replaced by 'key_name' in the CCs code.

Later on on can then change the column name of the 'key' column in all
annotations tabels to 'key_name' transparently without downtime and
remove the view in a third step.

This change is a refactoring to streamline annotations and labels
and finish tasks proposed in
/decitions/0004-adding-key-prefix-to-annotations.md