Correct migration tests to not leave db in half migrated state
This change fixes the migration tests for the streamlining of annotation
keys. The previous test would leave the db in a half migrated state and
then following tests would fail due to the rest of the migrations beeing
missing.

It may make sense to move the begin and after part of the migration test
into a helper to make this a proper workflow for every migration test
since we must make sure we start the test before the migration we`d like
to test. Then just apply one single migration(and not all migrations),
then expect certain behaviour and lastly complete the migration to not
destroy the DB. This should be in a helper as currently it also requires
working down migrations which is not always the case.