Introduce shared context for migrations and add documentation
Previously a migration was only tested by reverting (running down
migrations) until being one migration before the one that shall be
tested, then executing the actual test in which all migrations run the
up part even of migrations that come after the migration one wants to
test.

This commit introduces the 'migration' shared context, a valuable
utility for Sequel migration testing. This new module reverts the
database schema to its version prior to a specific migration, provides a
directory with the migration file for running a given migration in a
test, and restores the schema to its full state after test completion to
avoid half-migrated database scenarios. Hereby only the up part of the
migration one wants to test is executed opposite to all available
migrations to construct a meaningful test case.

The accompanying updates to the documentation elaborate on the
distinctive operating nature of Sequel migration tests. Also extensive
documentation on Sequel migrations itself in the scope of the supported
dbs and cloud_controller_ng has been added to provide a good knowledge
base for writing highly sophisticated migrations that are resilient,
offer high level of consistency, are compatible with mysql and psql and
fast.

Constraints and restrictions on Sequel migrations and test writing have
also been detailed, stressing the need to exclude Cloud Controller code
in migrations and tests to maintain consistency in test outcomes and
behaviours.

Co-authored-by: Philipp Thun <philipp.thun@sap.com>