Make restructure_index migration idempotent and small transactions
This fixes a deadlock issue when running the migration
20230725110800_restructure_index_for_improved_pagination_order.rb

This migration migrated halve the DBs tables in a single transaction
because `up do` implicitly creates a transaction inside its whole
block. This has the side effect that queries with joins can lock some
tables while the transaction locks tables the select also waits on and
vice versa. Creating a deadlock situation. The migration in its previous
state needed some luck with lock.

This change fixes the scenario so no deadlock can occur by running the
migration for each table in an own transaction. So the whole transaction
just requires a lock for a single table which prevents any mutual
deadlock situations.

Additionally the migration was made idempotent as when such a deadlock
happens on postgres the DLLs also get transactionally reverted so the
migration can start over again. However this is not the case for mysql.
If the transaction is aborted due to a deadlock half of the tables
indices stay migrated the other half is not. The previous implementation
then would fail e.g. on a drop index due to the index being already
drop creating a situation where manual intervention is needed for mysql.
This now cannot happen anymore due to idempotency of the migration.

Co-authored-by: Katharina Przybill <katharina.przybill@sap.com>