add script to start docker container with db for unit tests
I don't work on a machine that has a local DB. When I started working on
readiness healthcheck stuff I spent 2+ hours just trying to get one
setup. In the end, I made this script from a similar one in
cf-networking-release. I don't promise that it works for every CAPI use
case (it only works with postgres for example), but it works great for
when a new person wants to run some tests and iterate quickly.

To use, run: `./scripts/docker-shell-with-started-db`