Fix a performance regression in /v3/(space|organization)/:guid/users
The endpoints for listing space and org users has a performance issue
when there is a large amount of roles in the roles tables and/or many
users exists in different role tables.

The query is constructed in a way that all role tables get unioned and
then a `select ... from users where user_id in (<RESULT OF UNION>)` is
executed. PSQL query planner took bad decitions when many duplicates
exist and the resulting union is large(observed at arount 1500 rows).
This changed the query times from around 10ms to 55seconds.

This change introduces a `select distinct` in the `in` clause so that no
duplicates exist which is beneficial in any case for this endpoint.
Now the query planner of PSQL query planner decides correctly.

Co-authored-by: Philipp Thun <philipp.thun@sap.com>