Extend rubocop and add rubocop_sequel cops
Rubocop configuration was not overhauled for quite some time and new
cops and recommended patterns do not get enabled by default.
This change overhauls the rubocop configuration to apply the latest
styleguides. Also it enables the use of rubocop_sequel which gives
usefull linting for using the sequel library for writing migrations and
also for usage in the applications codebase.