Use consistent hash structure for process stats
The ProcessStatsPresenter distinguishes between 'down' and 'found'
instances. For 'down' instances it expects the 'uptime' key in the given
hash itself, for 'found' instances, it is part of the 'stats' hash.
Although the ReporterMixin constructed 'down' instances as expected,
there were cases (e.g. placement errors) where a different hash was
created. This resulted in a JSON response containing "uptime":null.

With this change a consistent hash structure is used for all cases, i.e.
'uptime' is a key in 'stats'.