Improve draining (for Puma)
Introduce a 'final timeout' for nginx and cc: i.e. check every second if
the process got terminated; if the 'final timeout' is reached and the
process is still there, write an error log.

Without this additional timeout it was (in theory) possible that the
termination of cc was triggered although the forceful nginx shutdown has
not been finished. And the overall drain procedure could have been
completed even though cc was not fully terminated.

For cc the 'final timeout' is 20 seconds. This has been derived from the
shutdown timeouts for Puma workers (15s) and threads (10s). This value
should be fine for Thin as well.

For nginx the 'final timeout' is 10 seconds.

The explicit log message will make it possible to detect draining issues
that might have been unnoticed so far.