Fix: Polling stopped after failed service instance update last operation fetch
The Cloud Controller stopped polling the last operation state of a service instance update request
when the Service Broker was unavailable and a proxy returned a 502 for example.
According to the OSBAPI "Responses with any other status code SHOULD be interpreted as an error or invalid response. The Platform SHOULD continue polling until the Service Broker returns a valid response or the maximum polling duration is reached."

Added tests to all polling operations to make sure the behaviour is consistent.