Fix missing statsd config error on workers
Workers do not emit metrics (there is no statsd_injector process) and
thus don't need any statsd config. As only a single method of the
Diego::Messenger is using statsd_updater and prometheus_updater and this
method is not used by worker processes, these updaters must not be
initialized in the constructor.