Stop polling and raise on standard error (service_instance_delete)
Previously all standard errors were caught, the last operation was set to failed, and polling continued.
Now HTTPResponseErrors and the like are separately caught, in which case the last operation stays 'in progress'
and the polling continues. For all other StandardErrors the last operation will be set to 'failed' and polling stops.
This is consistent for all other service related polling mechanisms.