Put binding and job creation in one transaction
When the DB becomes unavailable after the service binding operation was stored in the DB, but before the delayed job was stored, the operation will be stuck in state 'create initial'.
The user cannot delete the binding nor retrigger the creation. An admin has to set the operation to state failed manually in the DB.

This fixes it in that way, that both the operation and the delayed job get created in one DB transaction.