Fix test exclusion mechanism of backwards compatibillity tests
This Change fixes the exclusions of certain rspecs that trigger schema migrations for the rake task `spec:without_migrate'.
This task is useful when wanting to run the tests agains a predefined db schema e.g. a newer one, to test backwards compatibility.

With this change rspec and rspec_parallel properly include tagged test suites if the NO_DB_MIGRATION environment variable is set to an arbitrary value.