Switch to "using" Zeitwerk Rails autoloader
This is a prerequisite for upgrading the Rails 7. We don't appear to use
the Rails autoloader, so none of the module/class names/file paths are
conventional, which breaks Zeitwerk. I didn't find an obvious way to
turn off the Rails autoloader, so I instead ignored the entire app
directory, which appears to work. There may be a better way.