Ensure uniqueness of labels and annotations
This commit includes both code change and a database migration, contributing to ensuring labels and annotations' uniqueness.
The code update inserts an empty string as prefix into the database when the prefix is empty. It also addresses the race condition issue where two concurrent updates on an object’s labels could result in the label/annotation being created twice in the DB. A uniqueness constraint error is now caught, and the find_or_create operation is retried to handle this issue.
The database migration:

- Modifies the default value of the prefix column to an empty string.
- Converts all NULL values in the table to an empty string.
- Restricts the column from accepting NULL values.
- Reduces the length of the 'key' column in annotation tables to a maximum of 63 characters,
  in accordance with the documentation and current API input sanitation practices.
- Identifies and removes all duplicates based on identical values of resource_id, key_prefix, and
  key (annotation tables)/key_name (label tables).
- Implements a unique key constraint on resource_id, key_prefix, and
  key (annotation tables)/key_name (label tables) to prevent new duplicates.

Note that this change heavily utilizes table locking and transactions to ensure the change is carried out consistently and with resilience.
Also the change in the datatype of the key_prefix column is essential as
as for a unique constraint NULL vales are allways different as NULL !=
NULL in SQL99. A unique constraint containing NULL values would be not
of any use since it wont provide proper uniqueness for all columns it
contains.

Co-authored-by: Katharina Przybill <katharina.przybill@sap.com>