Prepare introduction of additional foreign keys
In preparation of adding a foreign key constraint from
buildpack_lifecycle_data to builds, random GUIDs must not be used for
field build_guid in tests.

To prevent backwards compatibility tests turning red, a check in
build_model_spec.rb needs to be disabled (and changed together with the
introduction of DELETE CASCADE).