Introduce config option to disable statsd metrics (#3585)
This config option allows to send metrics only via prometheus on api
vms. The default is true (i.e. enabled). In the default case, metrics are
sent via statsd and prometheus.

clock_schema and deployment_updater_schema need the optional
config because it is read in dependeny_locator and will error if the
config parameter is not there.