Change delayed_jobs_reserve index
Add WHERE condition. Jobs that failed are only kept for debugging
purposes, they are not taken into account when reserving a job. The
corresponding SELECT statement [1] includes 'WHERE failed_at IS NULL',
so the corresponding index can omit failed jobs.

Partial indexes are not supported for MySQL, thus this change is only
for PostgreSQL.

[1] https://github.com/TalentBox/delayed_job_sequel/
      blob/938ff68e98b44bbc391805c672a74c5ee3aef08b/
      lib/delayed/backend/sequel.rb#L26