Use DISTINCT only when required
As part of PR #3620 [1] it has been validated that there is no overlap
between the different sources (UNIONed datasets). But as these single
datasets are created by JOINing several tables, there might be
duplicates depending on the cardinality of those relations. Thus the
overall DISTINCT is replaced by DISTINCTs added when JOINing a table
with a one-to-many relation. Calling '.distinct' multiple times does not
have any negative effect as this simply sets a flag on the Sequel
dataset.

This change should improve performance of requests to the following
endpoints whenever those JOINs are not needed (e.g. depending on request
parameters):
- /v3/service_offerings
- /v3/service_plans

Tests now ensure that there are:
- multiple plans per offering
- multiple service instances per plan
- multiple visibilities per plan
- multiple spaces per organization

[1] https://github.com/cloudfoundry/cloud_controller_ng/pull/3620