Simplify migration-shared-context
When errors are encountered during the setup of migration tests, especially when DB migrations are run up to the one which is currently being tested, it was not possible to set breakpoints in the migration file. This was because the migration-shared-context copied the currently tested migration into its own folder and also the currently tested migration + all newer migrations into another folder. So breakpoints do not work.
Copying the migration files for only applying the correct migrations is not necessary. For the path we can use the normal path to the migrations folder. Additionally, we can use `target` to specify the migration to which we want to migrate. The correct target is calculated by using the current migrations timestamp and the latest migration's timestamp. This works for both, the up and the down path.