Use DISTINCT only when UNIONing multiple datasets (#3808)
When UNIONing multiple datasets, invoke 'abstract' method distinct_union
which is implemented differently by the ServiceOfferingListFetcher and
the ServicePlanListFetcher. For fetched service plans we already know
that there is no overlap between the subqueries; for fetched services
there can be duplicates and thus a DISTINCT has to be added to the
query.