Prevent 500 on parallel object creation
While fixing #3899, we found further possible occurrences of 500 errors in
severel _create actions when doing parallel requests.
In this PR we want to address them by adding an around_save to the
corresponding model in which we catch Sequel::UniqueConstraintViolation
errors for the case when  ne POST bypasses the other one and inserts the record into the db after the first POST has passed the validate function but before the actual insert.