Replace `jobs_user_guid_index` with `jobs_user_guid_state_index`
When the jobs table is large and single users have also a large amount of jobs, which are mostly in state `COMPLETED`,
postgres might decide to not use the index on `user_guid` column and scan the table with other methods, which are much slower.
This commit removes the index on user_guid column and adds a new partial index on user_guid and state where state is either
`POLLING` or `PROCESSING`. This keeps the index much smaller and improves the query performance drastically.

MySQL does not support partial indexes.