Use tcp on port 3000 for puma when nginx socket is empty
The current dev setup uses nginx to forward requests to CC.
`devenv.sh` script sets the nginx_socket to empty string as socket
communication can be problematic depending on the OS and Docker setup.
In this case Thin would open a tcp connection to port 3000.
This change does the same for Puma.