Ensure `Delayed::Worker` references work with ThreadedWorker (#3978)
Delayed::Plugin registers plugins statically against the lifecycle attribute of Delayed::Worker. This is not correct when the ThreadedWorker is used. To overcome this issue we need to ensure that ThreadedWorker acts as Worker by using an alias:

Delayed.module_eval do
  remove_const(:Worker) if const_defined?(:Worker)
  const_set(:Worker, Delayed::ThreadedWorker)
end

This ensures that all static references to Delayed::Worker are still working with ThreadedWorker.

Additionally some minor changes have been made:

    Moved the ThreadedWorker class into the Delayed module
    Add unit test for delayed job plugins
    Run ThreadedWorker related tests isolated from other tests to avoid side effects