Introduce `thread_grace_period_seconds` & `num_threads` parameters
This change allows operators to set the number of threads to start
multiple delayed worker threads (ThreadedWorker).
Additionally the grace period can be configured after which the threads
will be killed.
Both are implemented as parameters in the jobs rake task.
The existing `number_of_worker_threads` in the api and worker schema was
removed.