Drop constraint quota_definitions_name_key (#3952)
* Drop constraint quota_definitions_name_key

On the name column in quota_defintions there were two unique
constraints/indexes. One unique index with name qd_name_index, which is
similar in postgres and mysql, and then an unique constraint on column
name with name in postgres=quota_definitions_name_key and name in
mysql=name.
Since the unique index already exists, we don't need the unique
constraint. Dropping it with this migration.

* Adjust migration to support mysql 5.7

---------

Co-authored-by: Sven Krieger <37476281+svkrieger@users.noreply.github.com>