Add prometheus metrics to cc workers (#4205)
* Publish cc-worker metrics

Implemented a small puma webserver, which will be started in a separate thread with the first worker process.
Using the exporter middleware provided by the prometheus client, the metrics stored in the registry will be published under `/metrics`.
All processes use the same DirectFileStore registry, so that metrics can be published in a single webserver.

* Turn off publishing metrics by default and make configurable

* Use worker process index as pid for metrics store

* Fix code style and unit tests

* Make metrics endpoint only available on localhost